+++ amgcl/backend/block_crs.hpp
@@ -79,7 +79,7 @@ struct bcrs {
     {
 #pragma omp parallel
         {
-            std::vector<long> marker(bcols, -1);
+            std::vector<int> marker(bcols, -1);
 
 #ifdef _OPENMP
             int nt  = omp_get_num_threads();
@@ -152,7 +152,7 @@ struct bcrs {
 template <typename real>
 struct block_crs {
     typedef real value_type;
-    typedef long index_type;
+    typedef int index_type;
 
     typedef bcrs<real, index_type, index_type> matrix;
     typedef typename builtin<real>::vector     vector;
