+++ examples/mpi/call_mpi_lib.cpp
@@ -11,7 +11,7 @@
 
 #include "domain_partition.hpp"
 
-double STDCALL constant_deflation(int, long, void*) {
+double STDCALL constant_deflation(int, int, void*) {
     return 1;
 }
 
@@ -28,35 +28,35 @@ int main(int argc, char *argv[]) {
     if (rank == 0)
         std::cout << "World size: " << size << std::endl;
 
-    const long n  = argc > 1 ? atoi(argv[1]) : 1024;
-    const long n2 = n * n;
+    const int n  = argc > 1 ? atoi(argv[1]) : 1024;
+    const int n2 = n * n;
 
     // Partition
-    boost::array<long, 2> lo = { {0, 0} };
-    boost::array<long, 2> hi = { {n - 1, n - 1} };
+    boost::array<int, 2> lo = { {0, 0} };
+    boost::array<int, 2> hi = { {n - 1, n - 1} };
 
     domain_partition<2> part(lo, hi, size);
-    long chunk = part.size( rank );
+    int chunk = part.size( rank );
 
-    std::vector<long> domain(size + 1);
+    std::vector<int> domain(size + 1);
     MPI_Allgather(&chunk, 1, MPI_LONG, &domain[1], 1, MPI_LONG, MPI_COMM_WORLD);
     boost::partial_sum(domain, domain.begin());
 
-    long chunk_start = domain[rank];
-    long chunk_end   = domain[rank + 1];
+    int chunk_start = domain[rank];
+    int chunk_end   = domain[rank + 1];
 
-    std::vector<long> renum(n2);
-    for(long j = 0, idx = 0; j < n; ++j) {
-        for(long i = 0; i < n; ++i, ++idx) {
-            boost::array<long, 2> p = {{i, j}};
-            std::pair<int,long> v = part.index(p);
+    std::vector<int> renum(n2);
+    for(int j = 0, idx = 0; j < n; ++j) {
+        for(int i = 0; i < n; ++i, ++idx) {
+            boost::array<int, 2> p = {{i, j}};
+            std::pair<int,int> v = part.index(p);
             renum[idx] = domain[v.first] + v.second;
         }
     }
 
     // Assemble
-    std::vector<long>   ptr;
-    std::vector<long>   col;
+    std::vector<int>   ptr;
+    std::vector<int>   col;
     std::vector<double> val;
     std::vector<double> rhs;
 
@@ -69,8 +69,8 @@ int main(int argc, char *argv[]) {
 
     const double hinv = (n - 1);
     const double h2i  = (n - 1) * (n - 1);
-    for(long j = 0, idx = 0; j < n; ++j) {
-        for(long i = 0; i < n; ++i, ++idx) {
+    for(int j = 0, idx = 0; j < n; ++j) {
+        for(int i = 0; i < n; ++i, ++idx) {
             if (renum[idx] < chunk_start || renum[idx] >= chunk_end) continue;
 
             if (j > 0)  {
@@ -139,7 +139,7 @@ int main(int argc, char *argv[]) {
             std::ofstream f("out.dat", std::ios::binary);
             int m = n2;
             f.write((char*)&m, sizeof(int));
-            for(long i = 0; i < n2; ++i)
+            for(int i = 0; i < n2; ++i)
                 f.write((char*)&X[renum[i]], sizeof(double));
         } else {
             MPI_Send(x.data(), chunk, MPI_DOUBLE, 0, 42, MPI_COMM_WORLD);
